requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
web3==6.11.3
ipfshttpclient==0.8.0a2
cryptography==41.0.7
python-magic==0.4.27
pillow==10.1.0
redis==5.0.1
celery==5.3.4
email-validator==2.1.0 